--- Source/MonoGame.Extended/NuclexGui/Input/IInputService.cs
@@ -1,17 +1,0 @@
-﻿using MonoGame.Extended.InputListeners;
-
-namespace MonoGame.Extended.NuclexGui.Input
-{
-    public interface IInputService
-    {
-        KeyboardListener KeyboardListener { get; }
-
-        MouseListener MouseListener { get; }
-
-        GamePadListener GamePadListener { get; }
-
-        TouchListener TouchListener { get; }
-
-        void Update();
-    }
-}
